ai-engine/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
scikit-learn==1.5.2
numpy==2.1.3
pandas==2.2.3
openai==1.57.2
python-dotenv==1.0.1
pydantic==2.10.3
gitpython==3.1.43
requests==2.31.0
